buildextend-live: support embedding an Ignition config in the s390x ISO
This patch will allow users to embed Ignition configs in s390x ISOs.
This narrows the live ISO gap a bit between s390x and other arches (but
notably, ISO kargs support is still missing). It also crucially unblocks
running live ISO tests.

s390x does not support layered initrds, so to do this, we allocate space
directly in the initramfs passed to `mk-s390img` for the Ignition
config. We then create a new `/coreos/igninfo.json` file in the ISO
containing the metadata on how to find the Ignition embed area. On
!s390x, where layered initrds are supported, we point at `ignition.img`,
which maintains current behaviour. On s390x, we point at the right
offset of the reserved space in the `cdboot.img`.

coreos-installer will look for this `igninfo.json` file and embed the
Ignition config in the area to which it points.