platform/qemu: strengthen data block device checks
On ppc64le and s390x, the Ignition config is passed as a block device.
In the `RemovePrimaryBlockDevice()` code, we need to make sure we don't
inadvertently select the `ignition` block device as the new boot device.

More generally, the reliance on the `virtio-backend` string is weak.
Instead, use a more specific device name for the real block devices when
preparing the QEMU arguments, and filter on that.